+++ starkbank/merchantpurchase/__merchantpurchase.py
@@ -13,7 +13,7 @@ class MerchantPurchase(Resource):
                  billing_country_code=None, billing_city=None,billing_state_code=None, billing_street_line_1=None,
                  billing_street_line_2=None, billing_zip_code=None, metadata=None, card_ending=None, soft_descriptor=None,
                  challenge_mode=None, challenge_url=None, created=None, currency_code=None, end_to_end_id=None,
-                 fee=None, network=None, source=None, status=None, tags=None, updated=None):
+                 fee=None, network=None, source=None, status=None, tags=None, updated=None, confirmation_mode=None):
         Resource.__init__(self, id=id)
 
         self.amount = amount
@@ -45,6 +45,7 @@ class MerchantPurchase(Resource):
         self.source = source
         self.status = status
         self.tags = tags
+        self.confirmation_mode = confirmation_mode
         self.created = check_datetime(created)
         self.updated = check_datetime(updated)
 
@@ -96,3 +97,6 @@ def update(id, status=None, amount=None, user=None):
     }
     return rest.patch_id(resource=_resource, id=id, user=user, payload=payload)
 
+
+def delete(id, user=None):
+    return rest.delete_id(resource=_resource, id=id, user=user)